pyproject.toml
[tool.poetry]
name = "cloud_services"
version = "0.1.0"
description = ""
authors = ["Oleksandr Kudryavtsev <14zsoddenu@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^3.2.7"
django-environ = "^0.6.0"
django-ninja = "^0.14.0"
loguru = "^0.5.3"
psycopg2 = "^2.9.1"
djoser = "^2.1.0"
djangorestframework-simplejwt = "^4.8.0"
gunicorn = "^20.1.0"

[tool.poetry.dev-dependencies]
black = "^21.8b0"
pytest = "^6.2.5"
pytest-cov = "^2.12.1"
pytest-django = "^4.4.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120
exclude = '''
/(
  | venv
  | backend/models/__init__.py
  | migrations
)/
'''